+++ src/com/github/mjkuranda/retrographicsmanager/RetroGraphicsManager.java
@@ -6,22 +6,47 @@ import com.github.mjkuranda.retrographicsmanager.commandsystem.commands.Command;
 import java.util.Scanner;
 
 public class RetroGraphicsManager {
-    public static final String VERSION = "3.0.0";
-    public static final String RELEASE = "2023-05-11";
+    public static final String VERSION = "3.1.0";
+    public static final String RELEASE = "2023-05-17";
 
-    private boolean isRunning;
+    private static RetroGraphicsManager app;
+    private final ApplicationManager manager;
 
-    public RetroGraphicsManager() {
-        isRunning = false;
+    public static RetroGraphicsManager getInstance() {
+        if (app == null) {
+            app = new RetroGraphicsManager();
+        }
+
+        return app;
+    }
+
+    private RetroGraphicsManager() {
+        manager = new ApplicationManager();
+    }
+
+    private class ApplicationManager {
+        private boolean isRunning;
+
+        ApplicationManager() {
+            this.isRunning = false;
+        }
+
+        void setRunning(boolean isRunning) {
+            this.isRunning = isRunning;
+        }
+
+        boolean isRunning() {
+            return isRunning;
+        }
     }
 
     public void start() {
-        isRunning = true;
+        manager.setRunning(true);
         run();
     }
 
-    private void stop() {
-        isRunning = false;
+    public void stop() {
+        manager.setRunning(false);
     }
 
     private void run() {
@@ -33,39 +58,17 @@ public class RetroGraphicsManager {
         Scanner scan = new Scanner(System.in);
         String[] lineArgs;
 
-        while (isRunning) {
+        while (manager.isRunning()) {
             System.out.print("> ");
             lineArgs = scan.nextLine().split(" ");
-
-            if (ifTerminate(lineArgs)) {
-                stop();
-
-                continue;
-            }
-
             executeCommand(lineArgs);
         }
 
         System.out.println("Exit");
     }
 
-    private boolean ifTerminate(String[] lineArgs) {
-        if (lineArgs.length == 0) {
-            return false;
-        }
-
-        return switch(lineArgs[0]) {
-            case "exit", "end", "term", "terminate" -> true;
-            default -> false;
-        };
-    }
-
     private void executeCommand(String lineArgs[]) {
         Command c = CommandFactory.get(lineArgs);
-
-        if (c == null) {
-            return;
-        }
 
         if (!c.isValid()) {
             return;